requirements.txt
python3
py-cord
python-dotenv